requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
django-jazzmin
drf-spectacular>=0.28.0

